# src/main/java/com/ecommerce/complaints/controller/api/AiAnalysisApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.0.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.ecommerce.complaints.controller.api;

import com.ecommerce.complaints.model.vto.ComplaintAnalysisVTO;
import com.ecommerce.complaints.model.vto.ComplaintResponseVTO;
import com.ecommerce.complaints.model.vto.ErrorVTO;
import com.ecommerce.complaints.model.vto.ResponseGenerationRequestDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;

import java.io.IOException;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Tag(name = "AI Analysis", description = "the AI Analysis API")
public interface AiAnalysisApi {

    /**
     * POST /complaints/{id}/analyze : Perform detailed AI analysis on complaint
     *
     * @param id  (required)
     * @return  (status code 200)
     *         or Resource not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "analyzeComplaint",
        summary = "Perform detailed AI analysis on complaint",
        tags = { "AI Analysis" },
        responses = {
            @ApiResponse(responseCode = "200", description = "", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComplaintAnalysisVTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/complaints/{id}/analyze",
        produces = { "application/json" }
    )
    ResponseEntity<ComplaintAnalysisVTO> analyzeComplaint(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * POST /complaints/{id}/generate-response : Generate AI-powered response for complaint
     *
     * @param id  (required)
     * @param responseGenerationRequestDTO  (optional)
     * @return  (status code 200)
     *         or Resource not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "generateResponse",
        summary = "Generate AI-powered response for complaint",
        tags = { "AI Analysis" },
        responses = {
            @ApiResponse(responseCode = "200", description = "", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComplaintResponseVTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/complaints/{id}/generate-response",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<ComplaintResponseVTO> generateResponse(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "ResponseGenerationRequestDTO", description = "") @Valid @RequestBody(required = false) ResponseGenerationRequestDTO responseGenerationRequestDTO
    ) throws IOException;

}