# src/main/java/com/ecommerce/complaints/controller/api/ComplaintsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.0.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.ecommerce.complaints.controller.api;

import com.ecommerce.complaints.model.vto.ComplaintCreateDTO;
import com.ecommerce.complaints.model.vto.ComplaintListVTO;
import com.ecommerce.complaints.model.vto.ComplaintUpdateDTO;
import com.ecommerce.complaints.model.vto.ComplaintVTO;
import com.ecommerce.complaints.model.vto.ErrorVTO;
import com.ecommerce.complaints.model.vto.UpdateComplaintStatusRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Tag(name = "Complaints", description = "the Complaints API")
public interface ComplaintsApi {

    /**
     * POST /complaints : Create new complaint with auto-classification
     *
     * @param complaintCreateDTO  (required)
     * @return  (status code 201)
     *         or Bad request (status code 400)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "createComplaint",
        summary = "Create new complaint with auto-classification",
        tags = { "Complaints" },
        responses = {
            @ApiResponse(responseCode = "201", description = "", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComplaintVTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/complaints",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<ComplaintVTO> createComplaint(
        @Parameter(name = "ComplaintCreateDTO", description = "", required = true) @Valid @RequestBody ComplaintCreateDTO complaintCreateDTO
    );


    /**
     * DELETE /complaints/{id} : Delete a complaint by ID
     *
     * @param id  (required)
     * @return Complaint deleted successfully (status code 204)
     *         or Resource not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "deleteComplaint",
        summary = "Delete a complaint by ID",
        tags = { "Complaints" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Complaint deleted successfully"),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/complaints/{id}",
        produces = { "application/json" }
    )
    ResponseEntity<Void> deleteComplaint(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /complaints/{id} : Get single complaint by ID
     *
     * @param id  (required)
     * @return  (status code 200)
     *         or Resource not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getComplaintById",
        summary = "Get single complaint by ID",
        tags = { "Complaints" },
        responses = {
            @ApiResponse(responseCode = "200", description = "", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComplaintVTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/complaints/{id}",
        produces = { "application/json" }
    )
    ResponseEntity<ComplaintVTO> getComplaintById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /complaints : List complaints with filters and pagination
     *
     * @param status  (optional)
     * @param category  (optional)
     * @param priority  (optional)
     * @param sentiment  (optional)
     * @param page Page number (zero-based) (optional, default to 0)
     * @param size Page size (max 200) (optional, default to 20)
     * @param sortBy Field to sort by (optional, default to createdAt)
     * @param sortDirection Sort direction (optional, default to DESC)
     * @return Paginated list of complaints (status code 200)
     *         or Bad request (status code 400)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "listComplaints",
        summary = "List complaints with filters and pagination",
        tags = { "Complaints" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Paginated list of complaints", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComplaintListVTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/complaints",
        produces = { "application/json" }
    )
    ResponseEntity<ComplaintListVTO> listComplaints(
        @Parameter(name = "status", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "status", required = false) com.ecommerce.complaints.model.enums.ComplaintStatus status,
        @Parameter(name = "category", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "category", required = false) com.ecommerce.complaints.model.enums.ComplaintCategory category,
        @Parameter(name = "priority", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "priority", required = false) com.ecommerce.complaints.model.enums.Priority priority,
        @Parameter(name = "sentiment", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sentiment", required = false) com.ecommerce.complaints.model.enums.Sentiment sentiment,
        @Min(0) @Parameter(name = "page", description = "Page number (zero-based)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Min(1) @Max(200) @Parameter(name = "size", description = "Page size (max 200)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size,
        @Parameter(name = "sortBy", description = "Field to sort by", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sortBy", required = false, defaultValue = "createdAt") String sortBy,
        @Parameter(name = "sortDirection", description = "Sort direction", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sortDirection", required = false, defaultValue = "DESC") String sortDirection
    );


    /**
     * PUT /complaints/{id} : Update complaint details
     *
     * @param id  (required)
     * @param complaintUpdateDTO  (required)
     * @return  (status code 200)
     *         or Bad request (status code 400)
     *         or Resource not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "updateComplaint",
        summary = "Update complaint details",
        tags = { "Complaints" },
        responses = {
            @ApiResponse(responseCode = "200", description = "", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComplaintVTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/complaints/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<ComplaintVTO> updateComplaint(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "ComplaintUpdateDTO", description = "", required = true) @Valid @RequestBody ComplaintUpdateDTO complaintUpdateDTO
    );


    /**
     * PUT /complaints/{id}/status : Update complaint status
     *
     * @param id  (required)
     * @param updateComplaintStatusRequest  (required)
     * @return  (status code 200)
     *         or Bad request (status code 400)
     *         or Resource not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "updateComplaintStatus",
        summary = "Update complaint status",
        tags = { "Complaints" },
        responses = {
            @ApiResponse(responseCode = "200", description = "", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComplaintVTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/complaints/{id}/status",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<ComplaintVTO> updateComplaintStatus(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "UpdateComplaintStatusRequest", description = "", required = true) @Valid @RequestBody UpdateComplaintStatusRequest updateComplaintStatusRequest
    );

}