# src/main/java/com/ecommerce/complaints/swagger/swagger.yaml
openapi: 3.0.3
info:
  version: "1.0"
  title: "Complaint Analysis Service"
  contact:
    email: "zainmahmoud719@gmail.com"
    name: "MahmoudZain"

servers:
  - url: http://localhost:9000

tags:
  - name: Complaints
  - name: AI Analysis

paths:
  /complaints:
    post:
      tags: [Complaints]
      operationId: createComplaint
      summary: Create new complaint with auto-classification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplaintCreateDTO'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintVTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags: [ Complaints ]
      operationId: listComplaints
      summary: List complaints with filters and pagination
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ComplaintStatus'
        - name: category
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ComplaintCategory'
        - name: priority
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Priority'
        - name: sentiment
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Sentiment'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page number (zero-based)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
          description: Page size
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [ id, createdAt, updatedAt, status, priority, category, sentiment, subject ]
            default: createdAt
          description: Field to sort by
        - name: sortDirection
          in: query
          required: false
          schema:
            type: string
            enum: [ ASC, DESC ]
            default: DESC
          description: Sort direction
      responses:
        '200':
          description: Paginated list of complaints
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintListVTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'


  /complaints/{id}:
    get:
      tags: [Complaints]
      operationId: getComplaintById
      summary: Get single complaint by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintVTO'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Complaints]
      operationId: updateComplaint
      summary: Update complaint details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplaintUpdateDTO'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintVTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [ Complaints ]
      operationId: deleteComplaint
      summary: Delete a complaint by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Complaint deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /complaints/{id}/analyze:
    post:
      tags: [AI Analysis]
      operationId: analyzeComplaint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintAnalysisVTO'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /complaints/{id}/generate-response:
    post:
      tags: [AI Analysis]
      operationId: generateResponse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponseGenerationRequestDTO'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintResponseVTO'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /complaints/{id}/status:
    put:
      tags: [Complaints]
      operationId: updateComplaintStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateComplaintStatusRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintVTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorVTO'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorVTO'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorVTO'

  schemas:

    ComplaintCategory:
      type: string
      enum: [PRODUCT, SHIPPING, PAYMENT, REFUND, OTHER]

    ComplaintStatus:
      type: string
      enum: [NEW, IN_PROGRESS, RESOLVED, CLOSED]

    Sentiment:
      type: string
      enum: [POSITIVE, NEUTRAL, NEGATIVE]

    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH, URGENT]

    ComplaintCreateDTO:
      type: object
      required: [customerId, customerName, customerEmail, subject, description]
      properties:
        customerId:
          type: string
          minLength: 1
          maxLength: 100
        customerName:
          type: string
          minLength: 2
          maxLength: 100
        customerEmail:
          type: string
          format: email
          maxLength: 255
        subject:
          type: string
          minLength: 5
          maxLength: 255
        description:
          type: string
          minLength: 10
          maxLength: 5000

    ComplaintUpdateDTO:
      type: object
      properties:
        subject:
          type: string
          minLength: 5
          maxLength: 255
        description:
          type: string
          minLength: 10
          maxLength: 5000
        status:
          $ref: '#/components/schemas/ComplaintStatus'

    UpdateComplaintStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/ComplaintStatus'
        note:
          type: string
          maxLength: 500




    ResponseGenerationRequestDTO:
      type: object
      properties:
        tone:
          type: string
          enum: [professional, empathetic, apologetic, friendly]
          default: professional
        includeCompensation:
          type: boolean
          default: false
        customInstructions:
          type: string
          maxLength: 500

    ComplaintVTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
          format: email
        subject:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/ComplaintCategory'
        status:
          $ref: '#/components/schemas/ComplaintStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        sentiment:
          $ref: '#/components/schemas/Sentiment'
        aiAnalysis:
          type: string
        aiGeneratedResponse:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ComplaintListVTO:
      type: object
      properties:
        complaints:
          type: array
          items:
            $ref: '#/components/schemas/ComplaintVTO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        currentPage:
          type: integer
        pageSize:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean

    ComplaintAnalysisVTO:
      type: object
      properties:
        complaintId:
          type: integer
          format: int64
        category:
          $ref: '#/components/schemas/ComplaintCategory'
        categoryConfidence:
          type: number
          format: double
          minimum: 0.0
          maximum: 1.0
        sentiment:
          $ref: '#/components/schemas/Sentiment'
        sentimentScore:
          type: number
          format: double
          minimum: -1.0
          maximum: 1.0
        priority:
          $ref: '#/components/schemas/Priority'
        urgencyScore:
          type: number
          format: double
          minimum: 0.0
          maximum: 1.0
        urgencyReason:
          type: string
        keywords:
          type: array
          items:
            type: string
        summary:
          type: string
        recommendedActions:
          type: array
          items:
            type: string
        analyzedAt:
          type: string
          format: date-time

    ComplaintResponseVTO:
      type: object
      properties:
        complaintId:
          type: integer
          format: int64
        generatedResponse:
          type: string
        tone:
          type: string
        suggestedActions:
          type: array
          items:
            type: string
        similarComplaintsFound:
          type: integer
        confidenceScore:
          type: number
          format: double
        generatedAt:
          type: string
          format: date-time


    LookupItemVTO:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        isActive:
          type: boolean

    ErrorVTO:
      type: object
      properties:
        code:
          type: string
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        details:
          type: array
          items:
            type: string
        status:
          type: integer